07 - Document Q&A Chatbot - DataStax Astra DB/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
astrapy==1.1.0
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
bson==0.5.10
cachetools==5.3.3
cassandra-driver==3.29.1
cassio==0.1.7
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.6
datasets==2.19.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
emoji==2.11.1
faiss-cpu==1.8.0
filelock==3.14.0
filetype==1.2.0
frozenlist==1.4.1
fsspec==2024.3.1
geomet==0.2.1.post1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
groq==0.5.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.0
hyperframe==6.0.1
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langchain==0.1.4
langchain-astradb==0.3.0
langchain-community==0.0.16
langchain-core==0.1.52
langchain-groq==0.1.3
langchain-openai==0.0.3
langchain-text-splitters==0.0.1
langdetect==1.0.9
langsmith==0.1.57
llama-index==0.9.48
llama-parse==0.1.4
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
openai==1.28.1
orjson==3.10.3
packaging==23.2
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
pyarrow==16.0.0
pyarrow-hotfix==0.6
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
Pygments==2.18.0
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
ragstack-ai==0.7.0
rapidfuzz==3.9.0
referencing==0.35.1
regex==2024.5.10
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
streamlit==1.34.0
tabulate==0.9.0
tenacity==8.3.0
tiktoken==0.5.2
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
unstructured==0.10.30
urllib3==2.2.1
uuid6==2024.1.12
watchdog==4.0.0
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
